requirements.txt
duolingo-api==0.5.0
python-dotenv==0.13.0
python-telegram-bot==12.7
